requirements.txt
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
email-validator==2.2.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.8.0
limits==3.13.0
python-dotenv==1.0.1